- added support for sheet view options
